config: split compiler.go into cohesive same-package files (#2158)
Pure code-motion split of pkg/config/compiler.go (3050 LOC), the worst
splittable REFACTOR-tier offender in the #2158 audit heatmap and over the
3000 "must split before adding logic" line in the engineering-style
modularity discipline. Per the converged plan (docs/research/2158-file-
split/plan.md, section 5.2), the file is divided into four cohesive files,
all in `package config`:

  compiler.go (956)                  AST -> typed-struct compile core:
                                     the Err* retirement sentinels,
                                     compileOpts, CompileConfig*,
                                     compileConfigWithOpts,
                                     compileConfigForNode*, compileExpanded.

  compiler_validate_strict.go (947)  the validate*Strict commit-time
                                     hard-reject family (three-color
                                     policer, dataplane-type, policy
                                     scheduler / IPsec proposal / log
                                     profile references, routing-export
                                     #2144, policy match-address, the
                                     application-spec #2142, filter-protocol
                                     #2175, class-of-service strict
                                     validators) plus their dedicated
                                     helpers (routingRedistProtocolTokens,
                                     applicationsToValidateStrict /
                                     ApplicationsToValidateStrict,
                                     filterProtocolResolvable,
                                     policyMatchAddressError, the
                                     dataplaneType const block +
                                     effective/valid dataplane helpers,
                                     userspaceSynCookieProtectionActive,
                                     knownManagedProcessNames /
                                     isKnownProcessName).

  compiler_validate_warn.go (902)    ValidateConfig (non-fatal warnings)
                                     and its warning helpers (DDNS backend,
                                     routing-rule window, CoS
                                     oversubscription).

  compiler_applications.go (270)     compileApplications,
                                     parseApplicationTerms,
                                     normalizeProtocol, validatePortSpec,
                                     validateProtocol, nodeVal.

This is a tracking/janitorial refactor: no logic change, no exported-API
change, byte-identical runtime behavior. The split is purely
organizational -- Go resolves package-level symbols across files within a
package, so file placement does not affect compilation or behavior.

Behavior-preservation evidence:
  - go doc -all ./pkg/config/ is byte-identical before and after the
    split (proves zero exported-API change).
  - Reconstructing the original compiler.go from the four parts diffs to
    zero (modulo the section-seam blank lines), confirming move-only.
  - No init-order risk: the only package-level state relocated
    (routingRedistProtocolTokens, knownManagedProcessNames, the
    dataplaneType consts) are plain map/string literals with no
    side-effecting initializers; there is no init() in the package.
  - gofmt -l reports all four files clean.
  - go build ./..., go vet ./pkg/config/, and
    go test ./pkg/config/... ./pkg/daemon/... all pass, unchanged from
    master.
