+++ pkg/config/compiler.go
@@ -3,9 +3,6 @@ package config
 import (
 	"errors"
 	"fmt"
-	"net"
-	"sort"
-	"strconv"
 	"strings"
 )
 
@@ -953,2098 +950,4 @@ func compileExpanded(tree *ConfigTree, opts compileOpts) (*Config, error) {
 	// compile time — there is no value to nil out.
 
 	return cfg, nil
-}
-
-func validateThreeColorPolicersStrict(policers map[string]*ThreeColorPolicerConfig) error {
-	for name, pol := range policers {
-		if pol == nil {
-			continue
-		}
-		displayName := pol.Name
-		if displayName == "" {
-			displayName = name
-		}
-		if pol.SingleRateConfigured && pol.TwoRateConfigured {
-			return fmt.Errorf("firewall three-color-policer %q cannot configure both single-rate and two-rate", displayName)
-		}
-		if pol.ColorBlindConfigured && pol.ColorAwareConfigured {
-			return fmt.Errorf("firewall three-color-policer %q cannot configure both color-blind and color-aware", displayName)
-		}
-		if pol.CIR == 0 {
-			return fmt.Errorf("firewall three-color-policer %q requires positive committed-information-rate", displayName)
-		}
-		if pol.CBS == 0 {
-			return fmt.Errorf("firewall three-color-policer %q requires positive committed-burst-size", displayName)
-		}
-		if pol.PBS == 0 {
-			if pol.TwoRate {
-				return fmt.Errorf("firewall three-color-policer %q requires positive peak-burst-size", displayName)
-			}
-			return fmt.Errorf("firewall three-color-policer %q requires positive excess-burst-size", displayName)
-		}
-		if pol.TwoRate {
-			if pol.PIR == 0 {
-				return fmt.Errorf("firewall three-color-policer %q requires positive peak-information-rate", displayName)
-			}
-			if pol.PIR < pol.CIR {
-				return fmt.Errorf("firewall three-color-policer %q peak-information-rate must be >= committed-information-rate", displayName)
-			}
-			if pol.PBS < pol.CBS {
-				return fmt.Errorf("firewall three-color-policer %q peak-burst-size must be >= committed-burst-size", displayName)
-			}
-		}
-	}
-	return nil
-}
-
-// validateDataplaneTypeStrict rejects retired dataplane backends at
-// commit time. The parse path accepts `dataplane-type dpdk` as a
-// legal known value (see compileSystemDataplaneType +
-// validDataplaneType) so that `load merge` / `load override` of a
-// pre-retirement config does not syntax-error; this strict validator
-// is what tells the operator to migrate.
-//
-// Acceptance criterion in #1526 pins the verbatim message:
-//
-//	"the DPDK dataplane backend has been retired; use 'set system dataplane-type userspace' (see #1525)"
-//
-// Tests substring-match the phrase so the same assertion holds
-// whether the error is observed via `CompileConfig` directly,
-// via `Store.CommitCheck()` (which returns the raw error), or via
-// `Store.Commit()` (which wraps it as "commit check failed: ...").
-func validateDataplaneTypeStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	if cfg.System.DataplaneType == dataplaneTypeDPDK {
-		return ErrDPDKDataplaneRetired
-	}
-	if cfg.System.DataplaneType == dataplaneTypeEBPF {
-		return ErrEBPFDataplaneRetired
-	}
-	return nil
-}
-
-func validatePolicySchedulerReferencesStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	check := func(scope string, pol *Policy) error {
-		if pol == nil || pol.SchedulerName == "" {
-			return nil
-		}
-		if _, ok := cfg.Schedulers[pol.SchedulerName]; ok {
-			return nil
-		}
-		if scope != "" {
-			return fmt.Errorf("%s policy %q references undefined scheduler %q", scope, pol.Name, pol.SchedulerName)
-		}
-		return fmt.Errorf("policy %q references undefined scheduler %q", pol.Name, pol.SchedulerName)
-	}
-	for _, zpp := range cfg.Security.Policies {
-		if zpp == nil {
-			continue
-		}
-		for _, pol := range zpp.Policies {
-			if err := check("", pol); err != nil {
-				return err
-			}
-		}
-	}
-	for _, pol := range cfg.Security.GlobalPolicies {
-		if err := check("global", pol); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// validateIPsecPolicyProposalReferencesStrict hard-rejects an IPsec
-// (Phase 2) policy whose `proposals` reference does not resolve to a
-// defined IPsec proposal (#2073). resolveESPSettings (pkg/ipsec/ike.go)
-// resolves the policy's proposal ref, or falls back to the policy name
-// when no `proposals` leaf is given. When that reference dangles, the
-// renderer would otherwise fall through to `esp_proposals = default`,
-// silently substituting the operator's entire Phase-2 proposal set —
-// including any configured perfect-forward-secrecy DH group — with the
-// strongSwan default (which carries no required modp term). That is the
-// same silent-crypto-weakening class ValidateDHGroup closes for DH-group
-// leaves; this validator closes it for the policy→proposal cross-
-// reference.
-//
-// Rejected unconditionally (not only when PFSGroup > 0): a dangling
-// reference substitutes the whole proposal, not just PFS. Mirrors
-// validatePolicySchedulerReferencesStrict, which rejects any undefined
-// scheduler reference.
-//
-// On the tolerant load / peer-sync paths the call site downgrades this
-// to a warning (opts.lenientIPsecPolicyProposalRef) so an already-
-// persisted or peer-synced config still boots; the render-path safety
-// net in resolveESPSettings preserves the configured PFS group on that
-// boot rather than dropping it.
-func validateIPsecPolicyProposalReferencesStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	policies := cfg.Security.IPsec.Policies
-	proposals := cfg.Security.IPsec.Proposals
-	// Policies is a map (unordered); sort keys so the first-error
-	// commit-check message is deterministic across runs.
-	names := make([]string, 0, len(policies))
-	for name := range policies {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		pol := policies[name]
-		if pol == nil {
-			continue
-		}
-		propRef := pol.Proposals
-		explicitRef := propRef != ""
-		if !explicitRef {
-			// Mirror resolveESPSettings' policy-name fallback: a policy
-			// with no `proposals` leaf resolves against a proposal named
-			// after the policy itself.
-			propRef = pol.Name
-		}
-		if _, ok := proposals[propRef]; ok {
-			continue
-		}
-		if explicitRef {
-			return fmt.Errorf("ipsec policy %q references undefined ipsec proposal %q "+
-				"(the configured proposal set, including any perfect-forward-secrecy "+
-				"group, would be silently dropped to the strongSwan default)",
-				pol.Name, propRef)
-		}
-		// No explicit `proposals` leaf was given, so do not blame a
-		// phantom proposal named after the policy — describe the actual
-		// gap instead.
-		return fmt.Errorf("ipsec policy %q has no resolvable ipsec proposal "+
-			"(no `proposals` reference and no proposal named %q); the configured "+
-			"perfect-forward-secrecy group would be silently dropped — define a "+
-			"proposal or reference one", pol.Name, pol.Name)
-	}
-	return nil
-}
-
-// validateLogProfileStreamReferencesStrict hard-rejects a
-// `security log profile <name>` whose `stream-name` reference does not
-// resolve to a configured `security log stream` (#2008 H7). xpf routes
-// log events per stream (a Junos superset — every matching stream
-// receives the event), so a profile's `stream-name` designates the
-// stream that carries its events. A profile naming a stream that is not
-// configured routes to nowhere: the operator authored a log profile
-// whose target silently never fires. Before H7 the whole profile stanza
-// was dropped before compile, so the typo was invisible; now the
-// reference is validated.
-//
-// A profile with no `stream-name` is accepted: Junos permits a profile
-// that relies on the global routing inheritance, and there is nothing to
-// dangle. Only a non-empty `stream-name` that misses the stream map is
-// rejected.
-//
-// Note: compileLog only records a stream in Log.Streams when it has a
-// host (a host-less stream is not a real destination and is dropped by
-// the stream loop), so a profile referencing a host-less stream is
-// treated as a dangling reference — consistent with the stream's own
-// "must have a host to exist" semantics.
-//
-// On the tolerant load / peer-sync paths the call site downgrades this
-// to a warning (opts.lenientLogProfileStreamRef) so an already-persisted
-// config (older binaries dropped the stanza entirely) or a peer-synced
-// config still boots. Mirrors validateIPsecPolicyProposalReferencesStrict.
-func validateLogProfileStreamReferencesStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	profiles := cfg.Security.Log.Profiles
-	if len(profiles) == 0 {
-		return nil
-	}
-	streams := cfg.Security.Log.Streams
-	// Profiles is a map (unordered); sort keys so the first-error
-	// commit-check message is deterministic across runs.
-	names := make([]string, 0, len(profiles))
-	for name := range profiles {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		p := profiles[name]
-		if p == nil || p.StreamName == "" {
-			continue
-		}
-		if _, ok := streams[p.StreamName]; ok {
-			continue
-		}
-		return fmt.Errorf("security log profile %q references undefined "+
-			"log stream %q (the profile would route to nowhere — define "+
-			"the stream or fix the stream-name)", p.Name, p.StreamName)
-	}
-	return nil
-}
-
-// routingRedistProtocolTokens is the set of bare protocol keywords that an
-// OSPF/OSPFv3/BGP/IS-IS `export` (or a RIP `redistribute`) accepts in lieu
-// of a named policy-statement. resolveRedistribute (pkg/frr/policy_render.go)
-// emits a bare `redistribute <token>` for these. It mirrors
-// knownRedistProtocols there, plus Junos's `direct` spelling for FRR's
-// `connected`. Keep the two in sync: a token accepted here but unknown to
-// the renderer would emit a line FRR rejects; a token the renderer accepts
-// but missing here would be wrongly rejected at commit.
-var routingRedistProtocolTokens = map[string]bool{
-	"connected": true, "direct": true, "static": true, "kernel": true,
-	"ospf": true, "bgp": true, "rip": true, "isis": true,
-}
-
-// validateRoutingExportReferencesStrict hard-rejects a dynamic-protocol
-// `export` (OSPF / OSPFv3 / BGP / IS-IS), a RIP `redistribute`, a BGP
-// group/neighbor `export`, or a `routing-options forwarding-table export`
-// whose token resolves to neither a known redistribution protocol nor a
-// defined policy-statement (#2144).
-//
-// Without this gate a typo passes commit and reaches FRR render-time, where
-// it fails OPEN in three distinct ways:
-//
-//   - resolveRedistribute's fallback (policy_render.go) emits
-//     `redistribute <typo>` for any unknown token. FRR either rejects the
-//     line — failing the whole frr-reload (a single vtysh -f add-batch
-//     exits non-zero on any CMD_WARNING_CONFIG_FAILED) — or silently
-//     no-ops, so the intended redistribution never happens.
-//   - a BGP group/neighbor `export` renders `neighbor <addr> route-map
-//     <typo> out`. FRR resolves a route-map name with no definition to
-//     NULL, which it treats as permit-all — the outbound filter the
-//     operator wrote silently advertises EVERYTHING.
-//   - `forwarding-table export <typo>` is read by resolveECMP
-//     (config_render.go), which returns 0 max-paths when the policy is
-//     missing — silently DISABLING the expected ECMP / consistent-hash
-//     load balancing instead of rejecting the config.
-//
-// Protocol-token acceptance differs by site. A redistribute-backed export
-// (OSPF/OSPFv3/BGP/IS-IS export, RIP redistribute) legitimately names a
-// bare protocol (`export static`) OR a policy-statement, matching
-// resolveRedistribute. A BGP group/neighbor export and a forwarding-table
-// export render directly as a route-map / ECMP policy name, so only a
-// defined policy-statement is valid there — a protocol token would be a
-// dangling route-map / missing-policy reference, not a redistribute verb.
-//
-// On the tolerant load / peer-sync paths the call site downgrades this to a
-// warning (opts.lenientRoutingExportRef) so an already-persisted or
-// peer-synced config carrying the typo still boots (#1960
-// fail-closed-on-load class); the render-path fallbacks above keep it inert
-// or fail-open-on-an-already-committed-config exactly as before. Commit /
-// commit-check stay strict. Mirrors validateLogProfileStreamReferencesStrict.
-func validateRoutingExportReferencesStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	defined := func(name string) bool {
-		if cfg.PolicyOptions.PolicyStatements == nil {
-			return false
-		}
-		_, ok := cfg.PolicyOptions.PolicyStatements[name]
-		return ok
-	}
-
-	// checkRedist validates a redistribute-backed export list: each token
-	// must be a known protocol OR a defined policy-statement.
-	checkRedist := func(scope, proto string, exports []string) error {
-		for _, e := range exports {
-			if e == "" || routingRedistProtocolTokens[e] || defined(e) {
-				continue
-			}
-			return fmt.Errorf("%s%s export %q references neither a known "+
-				"redistribution protocol (connected/direct/static/kernel/"+
-				"ospf/bgp/rip/isis) nor a defined policy-statement — the "+
-				"FRR redistribute line would be rejected or silently no-op; "+
-				"define the policy-statement or fix the export name",
-				scope, proto, e)
-		}
-		return nil
-	}
-
-	// checkPolicyRef validates an export list that renders directly as a
-	// route-map / ECMP policy name: only a defined policy-statement is valid.
-	checkPolicyRef := func(detail, name string) error {
-		if name == "" || defined(name) {
-			return nil
-		}
-		return fmt.Errorf("%s references undefined policy-statement %q; %s",
-			detail, name, "define the policy-statement or fix the export name")
-	}
-
-	checkProtocols := func(scope string, ospf *OSPFConfig, ospfv3 *OSPFv3Config, bgp *BGPConfig, rip *RIPConfig, isis *ISISConfig) error {
-		if ospf != nil {
-			if err := checkRedist(scope, "protocols ospf", ospf.Export); err != nil {
-				return err
-			}
-		}
-		if ospfv3 != nil {
-			if err := checkRedist(scope, "protocols ospf3", ospfv3.Export); err != nil {
-				return err
-			}
-		}
-		if rip != nil {
-			if err := checkRedist(scope, "protocols rip", rip.Redistribute); err != nil {
-				return err
-			}
-		}
-		if isis != nil {
-			if err := checkRedist(scope, "protocols isis", isis.Export); err != nil {
-				return err
-			}
-		}
-		if bgp != nil {
-			if err := checkRedist(scope, "protocols bgp", bgp.Export); err != nil {
-				return err
-			}
-			// A BGP group/neighbor export renders `route-map <name> out`,
-			// so it must be a defined policy-statement (no protocol-token
-			// fallback). Sort neighbor addresses for a deterministic
-			// first-error message.
-			neighbors := append([]*BGPNeighbor(nil), bgp.Neighbors...)
-			sort.SliceStable(neighbors, func(i, j int) bool {
-				return neighbors[i].Address < neighbors[j].Address
-			})
-			for _, n := range neighbors {
-				if n == nil {
-					continue
-				}
-				for _, e := range n.Export {
-					detail := fmt.Sprintf("%sprotocols bgp neighbor %s export", scope, n.Address)
-					if n.GroupName != "" {
-						detail = fmt.Sprintf("%sprotocols bgp group %s neighbor %s export", scope, n.GroupName, n.Address)
-					}
-					if err := checkPolicyRef(detail, e); err != nil {
-						return err
-					}
-				}
-			}
-		}
-		return nil
-	}
-
-	// Top-level protocols.
-	if err := checkProtocols("", cfg.Protocols.OSPF, cfg.Protocols.OSPFv3, cfg.Protocols.BGP, cfg.Protocols.RIP, cfg.Protocols.ISIS); err != nil {
-		return err
-	}
-
-	// Per routing-instance protocols.
-	for _, ri := range cfg.RoutingInstances {
-		if ri == nil {
-			continue
-		}
-		scope := fmt.Sprintf("routing-instance %s ", ri.Name)
-		if err := checkProtocols(scope, ri.OSPF, ri.OSPFv3, ri.BGP, ri.RIP, ri.ISIS); err != nil {
-			return err
-		}
-	}
-
-	// forwarding-table export → resolveECMP (config_render.go). Renders
-	// directly as an ECMP policy lookup, so it must be a defined
-	// policy-statement; a missing one silently disables ECMP/consistent-hash.
-	if err := checkPolicyRef(
-		"routing-options forwarding-table export",
-		cfg.RoutingOptions.ForwardingTableExport,
-	); err != nil {
-		return fmt.Errorf("%s (the expected ECMP / consistent-hash "+
-			"load-balancing would be silently disabled)", err)
-	}
-
-	return nil
-}
-
-// validatePolicyMatchAddressesStrict hard-rejects a policy
-// source-address / destination-address token that is neither a known
-// address-book name (Address or AddressSet), the `any` keyword, nor a
-// parseable CIDR / bare IP (#2008). Such a token (a typo) reaches the
-// dataplane as an opaque string, fails CIDR/IP parsing in the Rust
-// literal parser, and is silently dropped to an empty set. Under
-// `*-address-excluded` inversion an empty set evaluates to MATCH-ALL —
-// a silent fail-open security bypass (a policy meant to exclude one
-// address ends up matching every address). Failing the typo at commit
-// turns the bypass into an operator-visible error.
-//
-// Legitimate forms accepted: address-book names, `any` (and the
-// family-scoped `any-ipv4` / `any-ipv6`, which compilePolicy already
-// normalizes to `0.0.0.0/0` / `::/0` and which parse as CIDRs anyway),
-// literal CIDRs, and bare IPv4 / IPv6 addresses. Junos address RANGES
-// are an address-book construct (expanded to /32s under the book) and
-// are referenced from a policy only by book NAME, so no range form
-// reaches this token list.
-func validatePolicyMatchAddressesStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	// Collect valid address-book names (Addresses + AddressSets).
-	bookNames := make(map[string]bool)
-	if ab := cfg.Security.AddressBook; ab != nil {
-		for name := range ab.Addresses {
-			bookNames[name] = true
-		}
-		for name := range ab.AddressSets {
-			bookNames[name] = true
-		}
-	}
-	validToken := func(tok string) bool {
-		switch tok {
-		case "", "any", "any-ipv4", "any-ipv6":
-			return true
-		}
-		if bookNames[tok] {
-			return true
-		}
-		if _, _, err := net.ParseCIDR(tok); err == nil {
-			return true
-		}
-		return net.ParseIP(tok) != nil
-	}
-	check := func(scope string, pol *Policy) error {
-		if pol == nil {
-			return nil
-		}
-		for _, addr := range pol.Match.SourceAddresses {
-			if !validToken(addr) {
-				return policyMatchAddressError(scope, pol.Name, "source-address", addr)
-			}
-		}
-		for _, addr := range pol.Match.DestinationAddresses {
-			if !validToken(addr) {
-				return policyMatchAddressError(scope, pol.Name, "destination-address", addr)
-			}
-		}
-		return nil
-	}
-	for _, zpp := range cfg.Security.Policies {
-		if zpp == nil {
-			continue
-		}
-		for _, pol := range zpp.Policies {
-			if err := check("", pol); err != nil {
-				return err
-			}
-		}
-	}
-	for _, pol := range cfg.Security.GlobalPolicies {
-		if err := check("global", pol); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// validateApplicationSpecsStrict hard-rejects a user-defined application
-// (`set applications application <name> ...`) whose destination-port /
-// source-port is malformed (not a valid numeric port, port range, or known
-// service name, out of 1..65535, or an inverted low>high range) or whose
-// protocol token is not a known name, a junos-*
-// alias, or a 0..255 number (#2142) — but ONLY for applications that are
-// actually REFERENCED by a security policy or a source/destination-NAT rule's
-// `match application` (#2187), or for ALL applications when
-// `services application-identification` is enabled (every app then compiles
-// into the app-id catalog). Such a spec is accepted by ValidateConfig as a
-// WARNING only; commit succeeds, the dataplane app-id compiler records the
-// AppID name and then `continue`s past the unparsable port (a never-match
-// AppID), and a policy referencing the application fails CLOSED on a permit
-// rule or falls through OPEN on a deny rule. Failing the spec at commit turns
-// that silent semantic break into an operator-visible error.
-//
-// The referenced-only scope is deliberate (the issue's explicit
-// referenced-vs-unreferenced distinction): an UNREFERENCED, app-id-disabled
-// application definition compiles into nothing the dataplane can match against,
-// so its malformed spec cannot break a live policy decision — it stays a
-// warning so an operator iterating on a not-yet-wired application library is
-// not blocked, and so existing configs that carry an unreferenced app with a
-// port form the policy matcher cannot represent (e.g. a `source-port 0-N`
-// range, which Rust parse_port_spec rejects on low==0) still commit. The moment
-// such an app is referenced by a policy, or app-id is turned on, the gate
-// engages.
-//
-// It reuses validatePortSpec and validateProtocol — the same config-layer
-// validators that produce the warning in ValidateConfig — so no new divergent
-// port/protocol table is introduced (the dataplane's #2124 capability gate is
-// the runtime backstop, and these validators are a superset of what it admits).
-// Iteration is sorted by application name so the first-reported error is
-// deterministic across runs (Go map order is randomized).
-func validateApplicationSpecsStrict(cfg *Config) error {
-	if cfg == nil || len(cfg.Applications.Applications) == 0 {
-		return nil
-	}
-	toCheck := applicationsToValidateStrict(cfg)
-	if len(toCheck) == 0 {
-		return nil
-	}
-	names := make([]string, 0, len(toCheck))
-	for name := range toCheck {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		app := cfg.Applications.Applications[name]
-		if app == nil {
-			continue
-		}
-		if err := validatePortSpec(app.DestinationPort); err != nil {
-			return fmt.Errorf("application %q: destination-port: %w", name, err)
-		}
-		if err := validatePortSpec(app.SourcePort); err != nil {
-			return fmt.Errorf("application %q: source-port: %w", name, err)
-		}
-		if app.Protocol != "" {
-			if err := validateProtocol(app.Protocol); err != nil {
-				return fmt.Errorf("application %q: %w", name, err)
-			}
-		}
-	}
-	return nil
-}
-
-// applicationsToValidateStrict returns the set of user-defined application
-// names whose port/protocol spec is validated as a hard COMMIT error rather
-// than a warning. That is every user application referenced (directly, or as a
-// member of a referenced application-set) by a zone-pair or global security
-// policy, OR by a source/destination-NAT rule's `match application` (#2187 — a
-// NAT term consumes the app's port/proto the same way a policy does, so a
-// malformed app referenced only by a NAT rule must reject too), plus — when
-// `services application-identification` is enabled — every user application
-// (app-id compiles them all into the catalog). It mirrors the policy-reference
-// walk in appid.CatalogNames; the logic is duplicated here because pkg/appid
-// imports pkg/config (so the compiler cannot call back into appid without an
-// import cycle). Predefined junos-* applications are never returned — they are
-// not in cfg.Applications.Applications and their specs are owned by the
-// predefined table, not the operator. Static NAT carries no application match,
-// so only source and destination NAT rule-sets are walked.
-func applicationsToValidateStrict(cfg *Config) map[string]struct{} {
-	out := make(map[string]struct{})
-	if cfg == nil {
-		return out
-	}
-	userApps := cfg.Applications.Applications
-	// app-id enabled: the catalog compiles every user application, so validate
-	// them all.
-	if cfg.Services.ApplicationIdentification {
-		for name := range userApps {
-			out[name] = struct{}{}
-		}
-		return out
-	}
-	addRef := func(appName string) {
-		if appName == "" || appName == "any" {
-			return
-		}
-		if set, isSet := cfg.Applications.ApplicationSets[appName]; isSet {
-			expanded, err := ExpandApplicationSet(appName, &cfg.Applications)
-			if err == nil {
-				for _, member := range expanded {
-					if _, isUser := userApps[member]; isUser {
-						out[member] = struct{}{}
-					}
-				}
-				return
-			}
-			// ExpandApplicationSet bails on the FIRST dangling/undefined or
-			// over-nested member, which would otherwise let a MALFORMED user app
-			// that is ALSO a direct member of the same set escape the strict gate
-			// (commit silently succeeds — the #2142 fail-closed-on-permit
-			// pathology, scoped to a set carrying a dangling member). A dangling
-			// member is a separate existing concern; it must not mask a malformed
-			// spec on a sibling member. Fall back to the set's DIRECT user-app
-			// members so each one that resolves is still hard-rejected at commit.
-			if set != nil {
-				for _, member := range set.Applications {
-					if _, isUser := userApps[member]; isUser {
-						out[member] = struct{}{}
-					}
-				}
-			}
-			return
-		}
-		if _, isUser := userApps[appName]; isUser {
-			out[appName] = struct{}{}
-		}
-	}
-	walk := func(policies []*Policy) {
-		for _, pol := range policies {
-			if pol == nil {
-				continue
-			}
-			for _, appName := range pol.Match.Applications {
-				addRef(appName)
-			}
-		}
-	}
-	for _, zpp := range cfg.Security.Policies {
-		if zpp == nil {
-			continue
-		}
-		walk(zpp.Policies)
-	}
-	walk(cfg.Security.GlobalPolicies)
-
-	// #2187: a source/destination-NAT rule with `match application <name>` also
-	// consumes the referenced app's port/proto (pkg/dataplane/userspace/nat.go
-	// appPortsFromSpec). A malformed spec there returns nil ports, so the NAT
-	// term silently never-matches (or over-matches on a degenerate proto) with
-	// no commit error — and a bad app referenced ONLY by a NAT rule (not by any
-	// policy) escaped both this commit gate (policy-only) and the #2124 runtime
-	// gate (policy-only). Collect NAT-rule app references the same way as policy
-	// references (single app or application-set) so they are hard-rejected at
-	// commit, lenient on load — identical wiring to the policy path. Static NAT
-	// is intentionally not walked: StaticNATRule has no application match
-	// (compileNATStatic parses only source/destination-address), so there is no
-	// app reference to validate there.
-	walkNATRules := func(rs *NATRuleSet) {
-		if rs == nil {
-			return
-		}
-		for _, rule := range rs.Rules {
-			if rule == nil {
-				continue
-			}
-			addRef(rule.Match.Application)
-		}
-	}
-	for _, rs := range cfg.Security.NAT.Source {
-		walkNATRules(rs)
-	}
-	if cfg.Security.NAT.Destination != nil {
-		for _, rs := range cfg.Security.NAT.Destination.RuleSets {
-			walkNATRules(rs)
-		}
-	}
-	return out
-}
-
-// ApplicationsToValidateStrict exposes the strict-validation reference set
-// (the user-app names the commit-time gate hard-rejects) for cross-checking
-// against appid.CatalogNames. The strict walk INLINE-duplicates CatalogNames's
-// policy-reference resolution because pkg/appid imports pkg/config (so the
-// compiler cannot call back into appid without a cycle). This accessor lets a
-// pkg/appid test assert the two walks agree on the user-app subset, so a future
-// change to CatalogNames's resolution cannot let the compiler copy drift
-// silently. It is a TEST seam, not a runtime coupling — production code uses the
-// unexported validateApplicationSpecsStrict directly.
-func ApplicationsToValidateStrict(cfg *Config) map[string]struct{} {
-	return applicationsToValidateStrict(cfg)
-}
-
-// validateFilterProtocolsStrict hard-rejects any firewall-filter term whose
-// `from protocol <token>` is not resolvable by the centralized protocol SSOT
-// (#2175) — neither a known protocol name, a junos-* alias, nor a 0..255
-// number. It walks every inet and inet6 filter and reports the first offending
-// family / filter / term / token (sorted by filter name, then by the term's
-// position, so the first-reported error is deterministic across runs).
-//
-// Resolution goes through filterProtocolResolvable, which INLINE-mirrors the
-// acceptance set of appid.ProtocolNumber. The compiler cannot call
-// appid.ProtocolNumber directly because pkg/appid imports pkg/config (an import
-// cycle) — the same constraint that forces validateApplicationSpecsStrict to
-// duplicate appid.CatalogNames's policy-reference walk (#2142). A pkg/appid
-// drift-guard test (TestFilterProtocolResolvableMatchesProtocolNumber) asserts
-// the two acceptance sets agree via the exported FilterProtocolResolvable
-// accessor, so a future change to appid.ProtocolNumber cannot let this copy
-// drift silently.
-//
-// The dataplane compiler (compileFirewallFilters → validateFilterProtocols)
-// keeps an identical check as defense-in-depth, but its error is swallowed by
-// the daemon (compileErrorMustAbortApply == false): this commit-check gate is
-// what makes the refusal operator-visible. On the tolerant load / peer-sync
-// path the caller downgrades the returned error to a warning (#1960 no-brick).
-func validateFilterProtocolsStrict(cfg *Config) error {
-	if cfg == nil {
-		return nil
-	}
-	check := func(family string, filters map[string]*FirewallFilter) error {
-		names := make([]string, 0, len(filters))
-		for name := range filters {
-			names = append(names, name)
-		}
-		sort.Strings(names)
-		for _, name := range names {
-			filter := filters[name]
-			if filter == nil {
-				continue
-			}
-			for _, term := range filter.Terms {
-				if term == nil || term.Protocol == "" {
-					continue
-				}
-				if !filterProtocolResolvable(term.Protocol) {
-					return fmt.Errorf(
-						"firewall family %s filter %q term %q: unknown protocol %q "+
-							"(use a protocol name such as tcp/udp/icmp/icmpv6/gre/esp/ah/"+
-							"sctp/ospf or a numeric value 0-255)",
-						family, name, term.Name, term.Protocol)
-				}
-			}
-		}
-		return nil
-	}
-	if err := check("inet", cfg.Firewall.FiltersInet); err != nil {
-		return err
-	}
-	return check("inet6", cfg.Firewall.FiltersInet6)
-}
-
-// filterProtocolResolvable reports whether a `from protocol <token>` is
-// representable: it INLINE-mirrors the acceptance set of
-// appid.ProtocolNumber's ok==true result (the #2124/#2175 SSOT). pkg/config
-// cannot import pkg/appid (import cycle: pkg/appid imports pkg/config), so the
-// known-name set is duplicated here and pinned by the pkg/appid drift-guard
-// test TestFilterProtocolResolvableMatchesProtocolNumber via the exported
-// FilterProtocolResolvable accessor.
-//
-// The acceptance set is intentionally TIGHTER than validateProtocol (used by
-// validateApplicationSpecsStrict): validateProtocol blanket-accepts ANY
-// "junos-" prefix, but appid.ProtocolNumber only resolves the specific
-// junos-* aliases below, so an unknown "junos-foobar" must be rejected here to
-// stay consistent with the dataplane SSOT — otherwise commit would pass while
-// the swallowed dataplane gate dropped the constraint.
-func filterProtocolResolvable(token string) bool {
-	switch strings.ToLower(strings.TrimSpace(token)) {
-	case "tcp", "junos-tcp-any",
-		"udp", "junos-udp-any",
-		"icmp", "junos-icmp-all", "junos-ping",
-		"icmpv6", "icmp6", "junos-icmp6-all", "junos-pingv6",
-		"gre", "junos-gre",
-		"ospf", "junos-ospf",
-		"junos-ip-in-ip", "junos-ipip", "ipip",
-		"egp",
-		"igmp",
-		"pim",
-		"ah",
-		"esp",
-		"sctp",
-		"vrrp":
-		return true
-	default:
-		// Numeric protocol number, including the deliberate "0" (HOPOPT).
-		if n, err := strconv.Atoi(strings.TrimSpace(token)); err == nil && n >= 0 && n < 256 {
-			return true
-		}
-		return false
-	}
-}
-
-// FilterProtocolResolvable exposes filterProtocolResolvable for the pkg/appid
-// drift-guard test (TestFilterProtocolResolvableMatchesProtocolNumber), which
-// asserts this acceptance set agrees with appid.ProtocolNumber's ok==true
-// result so the INLINE-duplicated table cannot drift from the SSOT silently. It
-// is a TEST seam, not a runtime coupling — production code uses the unexported
-// filterProtocolResolvable directly.
-func FilterProtocolResolvable(token string) bool {
-	return filterProtocolResolvable(token)
-}
-
-func policyMatchAddressError(scope, polName, field, addr string) error {
-	if scope != "" {
-		return fmt.Errorf(
-			"%s policy %q: %s %q is not a defined address-book entry, the `any` keyword, or a valid CIDR/IP address",
-			scope, polName, field, addr)
-	}
-	return fmt.Errorf(
-		"policy %q: %s %q is not a defined address-book entry, the `any` keyword, or a valid CIDR/IP address",
-		polName, field, addr)
-}
-
-func validateClassOfServiceStrict(cos *ClassOfServiceConfig) error {
-	if cos == nil {
-		return nil
-	}
-	for _, sched := range cos.Schedulers {
-		if sched == nil {
-			continue
-		}
-		if sched.EqualFlowEnforcement && (!sched.TransmitRateExact || sched.TransmitRateBytes == 0) {
-			return fmt.Errorf(
-				"class-of-service scheduler %q equal-flow-enforcement requires positive transmit-rate exact",
-				sched.Name)
-		}
-		if sched.EqualFlowEnforcement && sched.SurplusSharing {
-			return fmt.Errorf(
-				"class-of-service scheduler %q equal-flow-enforcement cannot be combined with surplus-sharing",
-				sched.Name)
-		}
-		// #1746: the schema enum validator catches bad values on the
-		// set path; re-check here so externally-assembled configs
-		// cannot smuggle an unknown policy to the dataplane (which
-		// would silently parse it as "slowest").
-		switch sched.EqualFlowTargetPolicy {
-		case "", "slowest", "mean", "ideal-share":
-		default:
-			return fmt.Errorf(
-				"class-of-service scheduler %q equal-flow-target-policy %q is not one of slowest | mean | ideal-share",
-				sched.Name, sched.EqualFlowTargetPolicy)
-		}
-		// Both buffer-size forms set simultaneously is ambiguous. The compiler
-		// always clears the unused field (see compiler_class_of_service.go
-		// buffer-size case), so this can only arise in constructed or
-		// externally-assembled configs. Reject early rather than silently
-		// applying the "byte-size wins" runtime preference.
-		if sched.BufferSizeBytes > 0 && sched.BufferSizePercent > 0 {
-			return fmt.Errorf(
-				"class-of-service scheduler %q has both buffer-size bytes (%d) "+
-					"and buffer-size percent (%.4g%%) set; use one form only",
-				sched.Name, sched.BufferSizeBytes, sched.BufferSizePercent)
-		}
-	}
-	// Aggregate percent check: Junos does not allow per-queue buffer
-	// allocations to exceed 100% of the interface's total buffer pool.
-	// Check each scheduler-map independently and reject overcommit here
-	// so the runtime never silently over-allocates. A sum of exactly
-	// 100% is permitted (full pool allocation).
-	//
-	// A small epsilon (1e-9) guards against accumulated IEEE 754 rounding
-	// when summing multiple float64 percent values: e.g. 33.33% * 3 may
-	// round to 99.99000000000001% rather than exactly 99.99%, so the check
-	// must not reject legitimate 100%-summing configs.
-	const maxTotalBufferPercent = 100.0
-	const bufferPercentEpsilon = 1e-9
-	for _, schedMap := range cos.SchedulerMaps {
-		if schedMap == nil {
-			continue
-		}
-		var totalPercent float64
-		for _, entry := range schedMap.Entries {
-			if entry == nil || entry.Scheduler == "" {
-				continue
-			}
-			sched, ok := cos.Schedulers[entry.Scheduler]
-			if !ok || sched == nil {
-				continue
-			}
-			totalPercent += sched.BufferSizePercent
-		}
-		if totalPercent > maxTotalBufferPercent+bufferPercentEpsilon {
-			return fmt.Errorf(
-				"class-of-service scheduler-map %q: "+
-					"sum of buffer-size percent across all schedulers is %.4g%% "+
-					"(must not exceed 100%%)",
-				schedMap.Name, totalPercent)
-		}
-	}
-	return nil
-}
-
-// #1830 (e): the #1733 MaxEqualFlowWorkers constant and
-// validateEqualFlowWorkerCapStrict were retired together with the Rust
-// MAX_WORKERS_SCRATCH cap they mirrored. The v8 lease rotation scratch
-// is now heap-sized to the true worker count, so equal-flow-enforcement
-// is supported at any configured worker count.
-
-// ValidateConfig performs non-fatal validation on a compiled config.
-// Returns warnings for unresolved references and operator-visible
-// compatibility/deprecation conditions.
-func ValidateConfig(cfg *Config) []string {
-	var warnings []string
-
-	// Note (#1476): the previous "ebpf is deprecated" warning was
-	// removed because `validateDataplaneTypeStrict` now hard-rejects
-	// `dataplane-type ebpf` at commit time with
-	// `ErrEBPFDataplaneRetired`. ValidateConfig is never reached for
-	// EBPF-typed configs after that gate; keeping the warning here
-	// would be dead code.
-
-	// #653: when `services application-identification` is enabled,
-	// emit a one-line warning at commit time so operators see what
-	// the knob actually does on xpf vs Junos vSRX. The runtime is
-	// port + protocol matching only — there is NO L7 DPI / signature
-	// engine. See `show services application-identification status`
-	// and docs/services-application-identification.md for the full
-	// contract.
-	if cfg.Services.ApplicationIdentification {
-		warnings = append(warnings,
-			"services application-identification is enabled, but xpf "+
-				"AppID is port+protocol catalog matching only — no L7 "+
-				"DPI / signature engine. Run `show services "+
-				"application-identification status` for the contract; "+
-				"see docs/services-application-identification.md.")
-	}
-
-	if userspaceSynCookieProtectionActive(cfg) &&
-		(cfg.System.RootAuthentication == nil ||
-			cfg.System.RootAuthentication.EncryptedPassword == "") {
-		warnings = append(warnings,
-			"active userspace-dp SYN-cookie screen profiles require "+
-				"system root-authentication encrypted-password material "+
-				"for the userspace cookie key; the userspace dataplane "+
-				"fails closed until it is set. Legacy eBPF SYN-cookie "+
-				"handling uses kernel helpers and is not affected by "+
-				"this warning.")
-	}
-
-	// #1944 §5.8: warn when a configured login user has no usable auth
-	// method — no ssh-* keys AND no usable encrypted-password (absent, or
-	// a bare lock sentinel which only locks the account). Mirrors the
-	// root-auth warning style above; directly addresses the "non-root
-	// operator cannot log in" bug class this issue closes.
-	if cfg.System.Login != nil {
-		for _, u := range cfg.System.Login.Users {
-			if u == nil || u.Name == "" || u.Name == "root" {
-				continue
-			}
-			// A usable password is a non-empty value that is neither a bare
-			// lock sentinel ("*"/"!"/"!!") NOR a locked-but-restorable form
-			// (any value beginning with "!", e.g. "!$6$salt$hash"). A
-			// leading "!" means the account cannot password-login until it
-			// is unlocked, so it does not count (Codex #1944 r1 Low).
-			pw := u.EncryptedPassword.Reveal()
-			usablePassword := pw != "" && pw != "*" && !strings.HasPrefix(pw, "!")
-			if len(u.SSHKeys) == 0 && !usablePassword {
-				warnings = append(warnings, fmt.Sprintf(
-					"login user %s has no usable authentication method: no "+
-						"ssh keys and no encrypted-password (a bare lock "+
-						"sentinel does not count) — this account cannot log "+
-						"in. Set `authentication encrypted-password` (hash "+
-						"from `openssl passwd -6`) or an ssh key.", u.Name))
-			}
-		}
-	}
-
-	// Collect valid zone names
-	zones := make(map[string]bool)
-	for name := range cfg.Security.Zones {
-		zones[name] = true
-	}
-
-	// Collect valid address-book entries
-	addrs := make(map[string]bool)
-	if ab := cfg.Security.AddressBook; ab != nil {
-		for name := range ab.Addresses {
-			addrs[name] = true
-		}
-		for name := range ab.AddressSets {
-			addrs[name] = true
-		}
-	}
-
-	// Collect valid applications
-	apps := make(map[string]bool)
-	for name := range cfg.Applications.Applications {
-		apps[name] = true
-	}
-	for name := range cfg.Applications.ApplicationSets {
-		apps[name] = true
-	}
-	// Built-in Junos application names
-	builtins := []string{"any", "junos-http", "junos-https", "junos-ssh", "junos-telnet",
-		"junos-dns-udp", "junos-dns-tcp", "junos-ping", "junos-icmp-all",
-		"junos-bgp", "junos-ospf", "junos-ntp", "junos-dhcp-relay",
-		"junos-ftp", "junos-smtp", "junos-icmp6-all", "junos-ike",
-		"junos-ipsec-nat-t", "junos-dhcp-client", "junos-dhcp-server",
-		"junos-snmp", "junos-syslog", "junos-traceroute", "junos-radius"}
-	for _, b := range builtins {
-		apps[b] = true
-	}
-
-	// Validate application port specs and protocols
-	for name, app := range cfg.Applications.Applications {
-		if err := validatePortSpec(app.DestinationPort); err != nil {
-			warnings = append(warnings, fmt.Sprintf("application %s: destination-port: %v", name, err))
-		}
-		if err := validatePortSpec(app.SourcePort); err != nil {
-			warnings = append(warnings, fmt.Sprintf("application %s: source-port: %v", name, err))
-		}
-		if app.Protocol != "" {
-			if err := validateProtocol(app.Protocol); err != nil {
-				warnings = append(warnings, fmt.Sprintf("application %s: %v", name, err))
-			}
-		}
-	}
-
-	// Validate policies
-	for _, zpp := range cfg.Security.Policies {
-		if zpp.FromZone != "any" && !zones[zpp.FromZone] {
-			warnings = append(warnings, fmt.Sprintf(
-				"policy from-zone %q: zone not defined", zpp.FromZone))
-		}
-		if zpp.ToZone != "any" && !zones[zpp.ToZone] {
-			warnings = append(warnings, fmt.Sprintf(
-				"policy to-zone %q: zone not defined", zpp.ToZone))
-		}
-		for _, p := range zpp.Policies {
-			for _, addr := range p.Match.SourceAddresses {
-				if addr != "any" && !addrs[addr] {
-					warnings = append(warnings, fmt.Sprintf(
-						"policy %q: source-address %q not in address-book", p.Name, addr))
-				}
-			}
-			for _, addr := range p.Match.DestinationAddresses {
-				if addr != "any" && !addrs[addr] {
-					warnings = append(warnings, fmt.Sprintf(
-						"policy %q: destination-address %q not in address-book", p.Name, addr))
-				}
-			}
-			for _, app := range p.Match.Applications {
-				if !apps[app] {
-					warnings = append(warnings, fmt.Sprintf(
-						"policy %q: application %q not defined", p.Name, app))
-				}
-			}
-		}
-	}
-
-	// Validate NAT zone references
-	for _, rs := range cfg.Security.NAT.Source {
-		if rs.FromZone != "" && !zones[rs.FromZone] {
-			warnings = append(warnings, fmt.Sprintf(
-				"source-nat ruleset %q: from-zone %q not defined", rs.Name, rs.FromZone))
-		}
-		if rs.ToZone != "" && !zones[rs.ToZone] {
-			warnings = append(warnings, fmt.Sprintf(
-				"source-nat ruleset %q: to-zone %q not defined", rs.Name, rs.ToZone))
-		}
-	}
-	// Static NAT rule-sets carry a `from zone` scope that the dataplane
-	// enforces on the inbound (DNAT) direction (static_nat.rs match_dnat:
-	// the entry is skipped unless its from_zone matches the ingress zone
-	// name exactly). A typo'd or undefined zone therefore yields a rule
-	// that silently never matches, with no other operator signal — mirror
-	// the source-NAT zone validation above so the divergence surfaces at
-	// commit (#2008 H15).
-	for _, rs := range cfg.Security.NAT.Static {
-		if rs == nil {
-			continue
-		}
-		if rs.FromZone != "" && !zones[rs.FromZone] {
-			warnings = append(warnings, fmt.Sprintf(
-				"static-nat ruleset %q: from-zone %q not defined", rs.Name, rs.FromZone))
-		}
-	}
-
-	// Validate screen references in zones
-	for name, zone := range cfg.Security.Zones {
-		if zone.ScreenProfile != "" {
-			if _, ok := cfg.Security.Screen[zone.ScreenProfile]; !ok {
-				warnings = append(warnings, fmt.Sprintf(
-					"zone %q: screen profile %q not defined", name, zone.ScreenProfile))
-			}
-		}
-	}
-
-	// Validate address-book entries have valid CIDR or IP formats
-	if ab := cfg.Security.AddressBook; ab != nil {
-		for name, entry := range ab.Addresses {
-			if entry.Value != "" {
-				if _, _, err := net.ParseCIDR(entry.Value); err != nil {
-					if net.ParseIP(entry.Value) == nil {
-						warnings = append(warnings, fmt.Sprintf(
-							"address-book %q: invalid address %q", name, entry.Value))
-					}
-				}
-			}
-		}
-		// Validate address-set members reference valid entries
-		for setName, as := range ab.AddressSets {
-			for _, m := range as.Addresses {
-				if !addrs[m] {
-					warnings = append(warnings, fmt.Sprintf(
-						"address-set %q: member %q not in address-book", setName, m))
-				}
-			}
-			for _, m := range as.AddressSets {
-				if !addrs[m] {
-					warnings = append(warnings, fmt.Sprintf(
-						"address-set %q: nested set %q not in address-book", setName, m))
-				}
-			}
-		}
-	}
-
-	// Validate static route destinations are valid CIDR
-	for _, sr := range cfg.RoutingOptions.StaticRoutes {
-		if sr.Destination != "" {
-			if _, _, err := net.ParseCIDR(sr.Destination); err != nil {
-				warnings = append(warnings, fmt.Sprintf(
-					"static route: invalid destination %q", sr.Destination))
-			}
-		}
-	}
-
-	// Validate DNAT pool references
-	if dnat := cfg.Security.NAT.Destination; dnat != nil {
-		for _, rs := range dnat.RuleSets {
-			for _, rule := range rs.Rules {
-				if rule.Then.PoolName != "" {
-					if _, ok := dnat.Pools[rule.Then.PoolName]; !ok {
-						warnings = append(warnings, fmt.Sprintf(
-							"destination-nat %q rule %q: pool %q not defined",
-							rs.Name, rule.Name, rule.Then.PoolName))
-					}
-				}
-			}
-		}
-	}
-
-	// Validate SNAT pool references
-	for _, rs := range cfg.Security.NAT.Source {
-		for _, rule := range rs.Rules {
-			if rule.Then.PoolName != "" {
-				if _, ok := cfg.Security.NAT.SourcePools[rule.Then.PoolName]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"source-nat %q rule %q: pool %q not defined",
-						rs.Name, rule.Name, rule.Then.PoolName))
-				}
-			}
-		}
-	}
-
-	// Validate zone interface references
-	configuredIfaces := make(map[string]bool)
-	for name := range cfg.Interfaces.Interfaces {
-		configuredIfaces[name] = true
-	}
-	for zoneName, zone := range cfg.Security.Zones {
-		for _, ifName := range zone.Interfaces {
-			// Strip unit suffix (e.g. "trust0.0" -> "trust0")
-			base := ifName
-			if idx := strings.Index(ifName, "."); idx > 0 {
-				base = ifName[:idx]
-			}
-			if !configuredIfaces[base] {
-				warnings = append(warnings, fmt.Sprintf(
-					"zone %q: interface %q not in interfaces config", zoneName, ifName))
-			}
-		}
-	}
-
-	// Validate scheduler references in policies
-	for _, zpp := range cfg.Security.Policies {
-		for _, p := range zpp.Policies {
-			if p.SchedulerName != "" {
-				if _, ok := cfg.Schedulers[p.SchedulerName]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"policy %q: scheduler %q not defined", p.Name, p.SchedulerName))
-				}
-			}
-		}
-	}
-	for _, p := range cfg.Security.GlobalPolicies {
-		if p.SchedulerName != "" {
-			if _, ok := cfg.Schedulers[p.SchedulerName]; !ok {
-				warnings = append(warnings, fmt.Sprintf(
-					"global policy %q: scheduler %q not defined", p.Name, p.SchedulerName))
-			}
-		}
-	}
-
-	// Validate routing-instance interface references
-	for _, ri := range cfg.RoutingInstances {
-		for _, ifName := range ri.Interfaces {
-			base := ifName
-			if idx := strings.Index(ifName, "."); idx > 0 {
-				base = ifName[:idx]
-			}
-			if !configuredIfaces[base] {
-				warnings = append(warnings, fmt.Sprintf(
-					"routing-instance %q: interface %q not in interfaces config",
-					ri.Name, ifName))
-			}
-		}
-	}
-
-	// Validate firewall filter references on interfaces
-	for ifName, ifc := range cfg.Interfaces.Interfaces {
-		for unitNum, unit := range ifc.Units {
-			if unit.FilterInputV4 != "" {
-				if _, ok := cfg.Firewall.FiltersInet[unit.FilterInputV4]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"interface %s unit %d: filter input %q not defined",
-						ifName, unitNum, unit.FilterInputV4))
-				}
-			}
-			if unit.FilterInputV6 != "" {
-				if _, ok := cfg.Firewall.FiltersInet6[unit.FilterInputV6]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"interface %s unit %d: filter input-v6 %q not defined",
-						ifName, unitNum, unit.FilterInputV6))
-				}
-			}
-			if unit.FilterOutputV4 != "" {
-				if _, ok := cfg.Firewall.FiltersInet[unit.FilterOutputV4]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"interface %s unit %d: filter output %q not defined",
-						ifName, unitNum, unit.FilterOutputV4))
-				}
-			}
-			if unit.FilterOutputV6 != "" {
-				if _, ok := cfg.Firewall.FiltersInet6[unit.FilterOutputV6]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"interface %s unit %d: filter output-v6 %q not defined",
-						ifName, unitNum, unit.FilterOutputV6))
-				}
-			}
-		}
-	}
-
-	// Validate chassis cluster fabric config
-	if cc := cfg.Chassis.Cluster; cc != nil {
-		// fabric1-interface without fabric1-peer-address (or vice versa) is incomplete
-		if (cc.Fabric1Interface != "") != (cc.Fabric1PeerAddress != "") {
-			warnings = append(warnings, "chassis cluster: fabric1-interface and fabric1-peer-address must both be set for dual-fabric")
-		}
-		// Check fabric interfaces are defined in interface config
-		for _, pair := range [][2]string{
-			{cc.FabricInterface, "fabric-interface"},
-			{cc.Fabric1Interface, "fabric1-interface"},
-		} {
-			ifName, label := pair[0], pair[1]
-			if ifName != "" {
-				if _, ok := cfg.Interfaces.Interfaces[ifName]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"chassis cluster %s %q: interface not defined", label, ifName))
-				}
-			}
-		}
-		// Check control interface is defined
-		if cc.ControlInterface != "" {
-			if _, ok := cfg.Interfaces.Interfaces[cc.ControlInterface]; !ok {
-				warnings = append(warnings, fmt.Sprintf(
-					"chassis cluster control-interface %q: interface not defined", cc.ControlInterface))
-			}
-		}
-		// Check fabric member interfaces don't overlap between fab0 and fab1
-		if cc.FabricInterface != "" && cc.Fabric1Interface != "" {
-			fab0Members := make(map[string]bool)
-			if f0 := cfg.Interfaces.Interfaces[cc.FabricInterface]; f0 != nil {
-				for _, m := range f0.FabricMembers {
-					fab0Members[m] = true
-				}
-			}
-			if f1 := cfg.Interfaces.Interfaces[cc.Fabric1Interface]; f1 != nil {
-				for _, m := range f1.FabricMembers {
-					if fab0Members[m] {
-						warnings = append(warnings, fmt.Sprintf(
-							"chassis cluster: fabric member %q shared between %s and %s",
-							m, cc.FabricInterface, cc.Fabric1Interface))
-					}
-				}
-			}
-		}
-	}
-
-	// Validate strict-vip-ownership requires VRRP (incompatible with no-reth-vrrp / private-rg-election)
-	if cc := cfg.Chassis.Cluster; cc != nil && (cc.NoRethVRRP || cc.PrivateRGElection) {
-		for _, rg := range cc.RedundancyGroups {
-			if rg.StrictVIPOwnership {
-				warnings = append(warnings, fmt.Sprintf(
-					"redundancy-group %d: strict-vip-ownership incompatible with no-reth-vrrp (no VRRP instances to gate on)", rg.ID))
-			}
-		}
-	}
-
-	// Warn if no-reth-vrrp set explicitly — redundant since private-rg-election is now default
-	if cc := cfg.Chassis.Cluster; cc != nil && cc.PrivateRGElection && cc.NoRethVRRP {
-		warnings = append(warnings, "chassis cluster: no-reth-vrrp is redundant (private-rg-election is the default)")
-	}
-
-	if cfg.System.PersistGroupsInheritance {
-		warnings = append(warnings, "system commit persist-groups-inheritance configured but group inheritance persistence is not implemented")
-	}
-
-	// #2008 H13 Stage 1: the leaf is now typed (schema + field) instead of
-	// being silently dropped, but the idle-yield dataplane runtime is not
-	// implemented — the userspace AF_XDP workers busy-poll. Warn so the
-	// operator knows the knob is accepted but currently has no effect.
-	if cfg.ForwardingOptions.AllowDataplaneSleep {
-		warnings = append(warnings, "forwarding-options allow-dataplane-sleep configured but is accepted-only — the userspace dataplane workers busy-poll and idle-yield is not yet implemented")
-	}
-
-	// #2078: the `security flow tcp-session` presence flags are typed and
-	// committed but the userspace AF_XDP dataplane enforces none of them
-	// today. no-syn-check / no-syn-check-in-tunnel would gate the
-	// session-create SYN check; rst-invalidate-session would tear a session
-	// down on RST; no-sequence-check (#2008 M9) would skip sequence-window
-	// validation. The dataplane session table is a pure 5-tuple flow entry
-	// with no TCP state machine and no sequence/window tracking, so there is
-	// nothing for any of these knobs to enforce or skip. This is an
-	// intentional, reviewed parity gap (see #2008 M9 and the RST design
-	// rationale in docs/active-active-new-connections.md); research #2078
-	// converged PLAN-KILL on enforcement. Warn so an operator who sets one
-	// of these is not silently misled into believing it has runtime effect.
-	if ts := cfg.Security.Flow.TCPSession; ts != nil {
-		var unenforced []string
-		if ts.NoSynCheck {
-			unenforced = append(unenforced, "no-syn-check")
-		}
-		if ts.NoSynCheckInTunnel {
-			unenforced = append(unenforced, "no-syn-check-in-tunnel")
-		}
-		if ts.RstInvalidateSession {
-			unenforced = append(unenforced, "rst-invalidate-session")
-		}
-		if ts.NoSequenceCheck {
-			unenforced = append(unenforced, "no-sequence-check")
-		}
-		if len(unenforced) > 0 {
-			warnings = append(warnings, fmt.Sprintf(
-				"security flow tcp-session %s configured but accepted-only — the userspace dataplane has no TCP state machine and does not enforce these knobs (config-only parity, #2078)",
-				strings.Join(unenforced, ", ")))
-		}
-	}
-
-	// #654: warn on `system processes X disable` for a process that
-	// bpfrx does not actually manage. Silently accepting the knob (as
-	// used to happen with e.g. `utmd disable` on vSRX) means the
-	// operator gets no signal that the setting is a no-op.
-	for _, proc := range cfg.System.DisabledProcesses {
-		if !isKnownProcessName(proc) {
-			warnings = append(warnings, fmt.Sprintf(
-				"system processes %q disable: bpfrx does not manage %q; setting has no runtime effect", proc, proc))
-		}
-	}
-
-	// #651: warn when archive-sites include inline `password`
-	// credentials. Runtime archival shells out to `scp` with
-	// `-o BatchMode=yes`, so the password is silently ignored and
-	// archival can fail unless matching SSH keys are already set up.
-	if cfg.System.Archival != nil {
-		for _, url := range cfg.System.Archival.ArchiveSitesWithPassword {
-			warnings = append(warnings, fmt.Sprintf(
-				"system archival archive-sites %q: inline password is accepted but ignored — archival uses scp BatchMode and relies on SSH keys, not passwords", url))
-		}
-	}
-
-	if cfg.System.Services != nil && cfg.System.Services.DNSProxyConfigured {
-		warnings = append(warnings, "system services dns dns-proxy configured but DNS proxy/forwarder runtime is not implemented")
-	}
-
-	// #1715: `system services dns` no longer selects a systemd-resolved
-	// owner runtime branch. xpf owns /etc/resolv.conf directly as a
-	// managed plain file and keeps resolved disabled+masked regardless of
-	// this stanza. Warn so an operator who set it expecting resolved is
-	// not surprised that resolved stays off.
-	if cfg.System.Services != nil && cfg.System.Services.DNSEnabled {
-		warnings = append(warnings, "system services dns: resolved-owner mode is not supported; xpf manages /etc/resolv.conf directly and keeps systemd-resolved disabled+masked")
-	}
-
-	if fm := cfg.Services.FlowMonitoring; fm != nil {
-		checkExtWarning := func(kind, name string, exts []string) {
-			for _, ext := range exts {
-				if ext == "app-id" {
-					warnings = append(warnings, fmt.Sprintf(
-						"flow-monitoring %s template %s: export-extension app-id configured but application data is not available in flow records", kind, name))
-				}
-			}
-		}
-		if fm.Version9 != nil {
-			for _, tmpl := range fm.Version9.Templates {
-				checkExtWarning("version9", tmpl.Name, tmpl.ExportExtensions)
-			}
-		}
-		if fm.VersionIPFIX != nil {
-			for _, tmpl := range fm.VersionIPFIX.Templates {
-				checkExtWarning("version-ipfix", tmpl.Name, tmpl.ExportExtensions)
-			}
-		}
-	}
-
-	if cos := cfg.ClassOfService; cos != nil {
-		warnedClassifierLossPriority := false
-		warnedRewriteLossPriority := false
-		for _, class := range cos.ForwardingClasses {
-			if class == nil {
-				continue
-			}
-			if class.Queue < 0 || class.Queue > 255 {
-				warnings = append(warnings, fmt.Sprintf(
-					"class-of-service forwarding-class %q uses out-of-range queue %d (expected 0..255)",
-					class.Name, class.Queue))
-			}
-		}
-		// #915: surplus-sharing is meaningful only on transmit-rate
-		// exact schedulers; warn-and-strip when set without exact so
-		// the runtime never sees the no-op flag (see #1183 lesson).
-		for _, sched := range cos.Schedulers {
-			if sched == nil {
-				continue
-			}
-			if sched.SurplusSharing && !sched.TransmitRateExact {
-				warnings = append(warnings, fmt.Sprintf(
-					"class-of-service scheduler %q surplus-sharing is meaningful only with transmit-rate exact; ignored",
-					sched.Name))
-				sched.SurplusSharing = false
-			}
-			// #1746: warn-not-strip. A policy without enforcement is a
-			// harmless no-op (the dataplane gates it on
-			// equal-flow-enforcement), but the operator should know it
-			// is inert.
-			if sched.EqualFlowTargetPolicy != "" && !sched.EqualFlowEnforcement {
-				warnings = append(warnings, fmt.Sprintf(
-					"class-of-service scheduler %q equal-flow-target-policy %q has no effect without equal-flow-enforcement",
-					sched.Name, sched.EqualFlowTargetPolicy))
-			}
-			// #1746: non-work-conserving cost warning. Clipping fast
-			// flows frees capacity that CANNOT reach slow flows on
-			// other workers, so these policies trade aggregate
-			// throughput for per-flow evenness (see
-			// docs/cos-traffic-shaping.md).
-			if sched.EqualFlowEnforcement {
-				switch sched.EqualFlowTargetPolicy {
-				case "slowest", "mean":
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service scheduler %q equal-flow-target-policy %s is non-work-conserving: it clips fast flows and reduces aggregate class throughput; it cannot lift slow flows",
-						sched.Name, sched.EqualFlowTargetPolicy))
-				}
-			}
-		}
-		for _, schedMap := range cos.SchedulerMaps {
-			if schedMap == nil {
-				continue
-			}
-			for className, entry := range schedMap.Entries {
-				if _, ok := cos.ForwardingClasses[className]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service scheduler-map %q references undefined forwarding-class %q",
-						schedMap.Name, className))
-				}
-				if entry == nil || entry.Scheduler == "" {
-					continue
-				}
-				if _, ok := cos.Schedulers[entry.Scheduler]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service scheduler-map %q references undefined scheduler %q",
-						schedMap.Name, entry.Scheduler))
-				}
-			}
-		}
-		for _, classifier := range cos.DSCPClassifiers {
-			if classifier == nil {
-				continue
-			}
-			for _, entry := range classifier.Entries {
-				if entry == nil || entry.ForwardingClass == "" {
-					continue
-				}
-				if _, ok := cos.ForwardingClasses[entry.ForwardingClass]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service dscp classifier %q references undefined forwarding-class %q",
-						classifier.Name, entry.ForwardingClass))
-				}
-				if entry.LossPriority != "" && !warnedClassifierLossPriority {
-					warnings = append(warnings, "class-of-service dscp/802.1p classifier loss-priority is accepted for compatibility but not yet enforced by the userspace dataplane")
-					warnedClassifierLossPriority = true
-				}
-			}
-		}
-		for _, classifier := range cos.IEEE8021Classifiers {
-			if classifier == nil {
-				continue
-			}
-			for _, entry := range classifier.Entries {
-				if entry == nil || entry.ForwardingClass == "" {
-					continue
-				}
-				if _, ok := cos.ForwardingClasses[entry.ForwardingClass]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service ieee-802.1 classifier %q references undefined forwarding-class %q",
-						classifier.Name, entry.ForwardingClass))
-				}
-				if entry.LossPriority != "" && !warnedClassifierLossPriority {
-					warnings = append(warnings, "class-of-service dscp/802.1p classifier loss-priority is accepted for compatibility but not yet enforced by the userspace dataplane")
-					warnedClassifierLossPriority = true
-				}
-			}
-		}
-		for _, rewriteRule := range cos.DSCPRewriteRules {
-			if rewriteRule == nil {
-				continue
-			}
-			for _, entry := range rewriteRule.Entries {
-				if entry == nil || entry.ForwardingClass == "" {
-					continue
-				}
-				if _, ok := cos.ForwardingClasses[entry.ForwardingClass]; !ok {
-					warnings = append(warnings, fmt.Sprintf(
-						"class-of-service dscp rewrite-rule %q references undefined forwarding-class %q",
-						rewriteRule.Name, entry.ForwardingClass))
-				}
-				if entry.LossPriority != "" && !warnedRewriteLossPriority {
-					warnings = append(warnings, "class-of-service dscp rewrite-rule loss-priority is accepted for compatibility but not yet enforced by the userspace dataplane")
-					warnedRewriteLossPriority = true
-				}
-			}
-		}
-		for _, iface := range cos.Interfaces {
-			if iface == nil {
-				continue
-			}
-			for _, unit := range iface.Units {
-				if unit == nil {
-					continue
-				}
-				if unit.SchedulerMap != "" {
-					if _, ok := cos.SchedulerMaps[unit.SchedulerMap]; !ok {
-						warnings = append(warnings, fmt.Sprintf(
-							"class-of-service interface %s unit %d references undefined scheduler-map %q",
-							iface.Name, unit.Unit, unit.SchedulerMap))
-					}
-				}
-				if unit.DSCPClassifier != "" {
-					if _, ok := cos.DSCPClassifiers[unit.DSCPClassifier]; !ok {
-						warnings = append(warnings, fmt.Sprintf(
-							"class-of-service interface %s unit %d references undefined dscp classifier %q",
-							iface.Name, unit.Unit, unit.DSCPClassifier))
-					}
-				}
-				if unit.IEEE8021Classifier != "" {
-					if _, ok := cos.IEEE8021Classifiers[unit.IEEE8021Classifier]; !ok {
-						warnings = append(warnings, fmt.Sprintf(
-							"class-of-service interface %s unit %d references undefined ieee-802.1 classifier %q",
-							iface.Name, unit.Unit, unit.IEEE8021Classifier))
-					}
-				}
-				if unit.DSCPRewriteRule != "" {
-					if _, ok := cos.DSCPRewriteRules[unit.DSCPRewriteRule]; !ok {
-						warnings = append(warnings, fmt.Sprintf(
-							"class-of-service interface %s unit %d references undefined dscp rewrite-rule %q",
-							iface.Name, unit.Unit, unit.DSCPRewriteRule))
-					}
-				}
-			}
-		}
-		hasCoSRuntimeConfig := len(cos.Interfaces) > 0 ||
-			len(cos.DSCPClassifiers) > 0 ||
-			len(cos.IEEE8021Classifiers) > 0 ||
-			len(cos.DSCPRewriteRules) > 0
-		if hasCoSRuntimeConfig && effectiveDataplaneType(cfg.System.DataplaneType) != dataplaneTypeUserspace {
-			warnings = append(warnings, "class-of-service shaping, classifier attachment, and dscp rewrite-rule attachment are only implemented in the userspace dataplane; configuration is accepted but will not take effect on this dataplane")
-		}
-
-		// #1614 A4: operator-visible warning when the sum of an
-		// interface unit's exact-class transmit-rates exceeds the
-		// unit's shaping-rate. Under oversubscription, every class
-		// will receive less than its configured rate; the visible
-		// distribution depends on the unit's oversubscription-policy.
-		warnings = append(warnings, validateCoSOversubscriptionWarnings(cos)...)
-	}
-
-	// #1706: the next-table and rib-group ip-rule reconcilers program
-	// into fixed 100-priority windows that their clear() passes scan.
-	// The applier hard-caps at the window boundary so out-of-range rules
-	// never leak, but a config that exceeds the window would be silently
-	// truncated at apply time. Surface the over-limit condition here at
-	// commit time so the operator sees it before applying.
-	warnings = append(warnings, validateRoutingRuleWindowWarnings(cfg)...)
-
-	// #1387: DHCP dynamic-DNS live-backend validation. Increment 2 wired the
-	// live RFC 2136 backend, so the increment-1 "no records are published"
-	// deferred-backend warning is retired. The warnings here flag a config
-	// that the now-live path cannot act on (enabled rfc2136 with no
-	// update-server), a still-deferred backend (kea-d2), and the now-consumed
-	// free-form leaves (update-server parseability, TSIG algorithm support).
-	// All are WARN-only (never an error) so a malformed inert value committed
-	// against increment 1 cannot brick a boot (plan §4.5 / §7 Q-C).
-	warnings = append(warnings, validateDDNSBackendWarnings(cfg)...)
-
-	return warnings
-}
-
-// validateDDNSBackendWarnings emits WARN-only commit-time messages for the
-// now-live DHCP dynamic-DNS backend (#1387 increment 2). It never returns
-// an error: the typed schema already accepts these leaves, and a stricter
-// HARD reject would brick a boot on a previously-inert malformed value
-// (plan §7 Q-C). The reconciler/backend degrade safely at runtime (an
-// unusable backend resolves to a no-op and counts a no-backend skip).
-func validateDDNSBackendWarnings(cfg *Config) []string {
-	d := cfg.System.DHCPServer.DynamicDNS
-	if d == nil || !d.Enabled {
-		return nil
-	}
-	var warnings []string
-
-	backend := d.Backend
-	if backend == "" {
-		backend = "rfc2136"
-	}
-	switch backend {
-	case "rfc2136":
-		if d.UpdateServer == "" {
-			warnings = append(warnings, "dhcp dynamic-dns is enabled with "+
-				"backend rfc2136 but no update-server is configured; no records "+
-				"will be published until an update-server is set")
-		} else if !ddnsUpdateServerParseable(d.UpdateServer) {
-			warnings = append(warnings, fmt.Sprintf("dhcp dynamic-dns "+
-				"update-server %q is not a valid host or host:port; the backend "+
-				"will fail to send updates", d.UpdateServer))
-		}
-		if d.TSIGKeyName != "" && !ddnsTSIGAlgorithmSupported(d.TSIGAlgorithm) {
-			warnings = append(warnings, fmt.Sprintf("dhcp dynamic-dns "+
-				"tsig-algorithm %q is not supported (use hmac-sha1, hmac-sha224, "+
-				"hmac-sha256, hmac-sha384, or hmac-sha512; hmac-md5 is rejected as "+
-				"insecure); the backend will fail to sign updates", d.TSIGAlgorithm))
-		}
-	case "kea-d2":
-		warnings = append(warnings, "dhcp dynamic-dns backend kea-d2 is "+
-			"reserved but not implemented (Kea D2 is not in the image); no "+
-			"records will be published with this backend")
-	}
-	return warnings
-}
-
-// ddnsUpdateServerParseable reports whether an update-server string is a
-// usable host or host:port (mirrors the backend's normalizeUpdateServer).
-func ddnsUpdateServerParseable(s string) bool {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return false
-	}
-	if _, _, err := net.SplitHostPort(s); err == nil {
-		return true
-	}
-	// No port: treat as a bare host (the backend attaches :53). Reject only
-	// when it is obviously not a host (e.g. embedded whitespace).
-	return !strings.ContainsAny(s, " \t")
-}
-
-// ddnsTSIGAlgorithmSupported reports whether a TSIG algorithm string is one
-// the backend can sign with (default hmac-sha256 when unset; hmac-md5
-// rejected). Mirrors ddns_rfc2136.canonicalTSIGAlgorithm without importing
-// the dhcpserver package into pkg/config.
-func ddnsTSIGAlgorithmSupported(algo string) bool {
-	a := strings.ToLower(strings.TrimSpace(algo))
-	a = strings.TrimSuffix(a, ".")
-	switch a {
-	case "", "hmac-sha1", "hmac-sha224", "hmac-sha256", "hmac-sha384", "hmac-sha512":
-		return true
-	default:
-		return false
-	}
-}
-
-// validateRoutingRuleWindowWarnings emits commit-time warnings when a
-// config would program more next-table or rib-group ip rules than the
-// applier's fixed priority window can hold (see pkg/routing/rules.go:
-// nextTableRulePriority window of 100, ribGroupRulePriority window of
-// 100 split into 50 v4+v6 pairs). The counts here are CONSERVATIVE
-// upper bounds computed from the same inputs the applier consumes —
-// they intentionally do NOT replicate the applier's exact skip/dedup
-// logic (unknown-instance skips, self-only rib-groups, duplicate source
-// tables), so they may warn slightly early but never miss a real
-// truncation. The runtime accepts the config; the apply-time cap is the
-// hard guard against the rule leak.
-func validateRoutingRuleWindowWarnings(cfg *Config) []string {
-	var warnings []string
-
-	// next-table: the applier feeds it the global inet + inet6 static
-	// routes (daemon_apply.go), counting those with a NextTable set.
-	const nextTableWindow = 100
-	nextTableRoutes := 0
-	for _, sr := range cfg.RoutingOptions.StaticRoutes {
-		if sr != nil && sr.NextTable != "" {
-			nextTableRoutes++
-		}
-	}
-	for _, sr := range cfg.RoutingOptions.Inet6StaticRoutes {
-		if sr != nil && sr.NextTable != "" {
-			nextTableRoutes++
-		}
-	}
-	if nextTableRoutes > nextTableWindow {
-		warnings = append(warnings, fmt.Sprintf(
-			"routing-options: %d static routes use next-table, but only %d can be "+
-				"programmed as ip rules; routes beyond the limit will be ignored at "+
-				"apply time. Reduce the number of next-table routes.",
-			nextTableRoutes, nextTableWindow))
-	}
-
-	// rib-group: the applier walks routing-instances and programs two ip
-	// rules (v4+v6) per source table that references an interface-routes
-	// rib-group. Window of 100 priorities fits 50 such tables.
-	const ribGroupTableLimit = 50
-	ribGroupInstances := 0
-	for _, inst := range cfg.RoutingInstances {
-		if inst == nil {
-			continue
-		}
-		if inst.InterfaceRoutesRibGroup != "" || inst.InterfaceRoutesRibGroupV6 != "" {
-			ribGroupInstances++
-		}
-	}
-	if ribGroupInstances > ribGroupTableLimit {
-		warnings = append(warnings, fmt.Sprintf(
-			"routing-options: %d routing-instances reference an interface-routes "+
-				"rib-group, but only %d leaking tables can be programmed as ip rules "+
-				"(two priorities each); instances beyond the limit will be ignored at "+
-				"apply time. Reduce the number of rib-group-leaking instances.",
-			ribGroupInstances, ribGroupTableLimit))
-	}
-
-	return warnings
-}
-
-// validateCoSOversubscriptionWarnings emits commit-time warnings for
-// every CoS interface unit whose sum of exact-class transmit rates
-// exceeds the unit's configured shaping-rate. Warnings are non-fatal;
-// the runtime accepts the config and the new
-// oversubscription-policy knob (#1614 A1) governs distribution.
-func validateCoSOversubscriptionWarnings(cos *ClassOfServiceConfig) []string {
-	var warnings []string
-	if cos == nil {
-		return warnings
-	}
-	for ifaceName, iface := range cos.Interfaces {
-		if iface == nil {
-			continue
-		}
-		for unitID, unit := range iface.Units {
-			if unit == nil || unit.ShapingRateBytes == 0 || unit.SchedulerMap == "" {
-				continue
-			}
-			schedMap, ok := cos.SchedulerMaps[unit.SchedulerMap]
-			if !ok || schedMap == nil {
-				continue
-			}
-			var sumExact uint64
-			for _, entry := range schedMap.Entries {
-				if entry == nil || entry.Scheduler == "" {
-					continue
-				}
-				sched, ok := cos.Schedulers[entry.Scheduler]
-				if !ok || sched == nil || !sched.TransmitRateExact {
-					continue
-				}
-				sumExact += sched.TransmitRateBytes
-			}
-			if sumExact <= unit.ShapingRateBytes {
-				continue
-			}
-			policyTail := "proportional (default): each class receives classRate × shaping / sumExact (current behaviour)"
-			if unit.OversubscriptionPolicy == "guarantee-rate" {
-				policyTail = fmt.Sprintf(
-					"guarantee-rate %g: small classes honoured to configured rate; larger classes share residual proportionally (see #1614)",
-					unit.OversubscriptionGuaranteeFraction,
-				)
-			}
-			warnings = append(warnings, fmt.Sprintf(
-				"class-of-service interfaces %s unit %d: sum of exact-class transmit-rates (%d B/s) exceeds shaping-rate (%d B/s); under oversubscription the configured oversubscription-policy=%s",
-				ifaceName, unitID, sumExact, unit.ShapingRateBytes, policyTail,
-			))
-		}
-	}
-	return warnings
-}
-
-const (
-	dataplaneTypeEBPF      = "ebpf"
-	dataplaneTypeDPDK      = "dpdk"
-	dataplaneTypeUserspace = "userspace"
-)
-
-func effectiveDataplaneType(dpType string) string {
-	if dpType == "" {
-		return dataplaneTypeUserspace
-	}
-	return dpType
-}
-
-func validDataplaneType(dpType string) bool {
-	switch dpType {
-	case dataplaneTypeEBPF, dataplaneTypeDPDK, dataplaneTypeUserspace:
-		return true
-	default:
-		return false
-	}
-}
-
-func userspaceSynCookieProtectionActive(cfg *Config) bool {
-	if cfg == nil || effectiveDataplaneType(cfg.System.DataplaneType) != dataplaneTypeUserspace ||
-		cfg.Security.Flow.SynFloodProtectionMode != "syn-cookie" {
-		return false
-	}
-	for _, zone := range cfg.Security.Zones {
-		if zone == nil || zone.ScreenProfile == "" {
-			continue
-		}
-		profile := cfg.Security.Screen[zone.ScreenProfile]
-		if profile != nil && profile.TCP.SynFlood != nil &&
-			profile.TCP.SynFlood.AttackThreshold > 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// knownManagedProcessNames is the set of Junos process names that bpfrx
-// actually honours when `system processes X disable` is configured.
-// The runtime sites hard-code their process name (not a table lookup):
-//   - pkg/daemon/daemon.go ~:715 — `isProcessDisabled(cfg, "snmpd")`
-//   - pkg/daemon/daemon_system.go ~:383 — `isProcessDisabled(cfg, "ntp")`
-//
-// This table mirrors those hard-codes for the purpose of the #654
-// validation warning. Any addition here MUST be paired with a matching
-// runtime gating site, or the warning will go quiet while the knob
-// remains a no-op.
-var knownManagedProcessNames = map[string]struct{}{
-	"snmpd": {},
-	"ntp":   {},
-}
-
-func isKnownProcessName(name string) bool {
-	_, ok := knownManagedProcessNames[name]
-	return ok
-}
-
-func compileApplications(node *Node, apps *ApplicationsConfig) error {
-	for _, inst := range namedInstances(node.FindChildren("application")) {
-		appName := inst.name
-		app := &Application{Name: appName}
-
-		var terms []*Application
-		for _, prop := range inst.node.Children {
-			switch prop.Name() {
-			case "protocol":
-				app.Protocol = nodeVal(prop)
-			case "destination-port":
-				app.DestinationPort = nodeVal(prop)
-			case "source-port":
-				app.SourcePort = nodeVal(prop)
-			case "inactivity-timeout", "timeout":
-				if v := nodeVal(prop); v != "" {
-					if n, err := strconv.Atoi(v); err == nil {
-						app.InactivityTimeout = n
-					}
-				}
-			case "alg":
-				app.ALG = nodeVal(prop)
-			case "description":
-				app.Description = nodeVal(prop)
-			case "term":
-				// Inline term: "term <name> [alg <a>] protocol <p> [source-port <sp>]
-				//               [destination-port <dp>] [inactivity-timeout <t>];"
-				if len(prop.Keys) < 2 {
-					continue
-				}
-				// Hierarchical: all values in prop.Keys (inline statement)
-				// Flat set: values split across prop.Keys and prop.Children
-				allKeys := prop.Keys[1:]
-				for _, c := range prop.Children {
-					allKeys = append(allKeys, c.Keys...)
-				}
-				termApps := parseApplicationTerms(appName, allKeys)
-				terms = append(terms, termApps...)
-			}
-		}
-
-		if len(terms) > 0 {
-			implicitSet := &ApplicationSet{Name: appName}
-			for _, t := range terms {
-				t.Description = app.Description
-				apps.Applications[t.Name] = t
-				implicitSet.Applications = append(implicitSet.Applications, t.Name)
-			}
-			apps.ApplicationSets[appName] = implicitSet
-		} else {
-			apps.Applications[appName] = app
-		}
-	}
-
-	for _, inst := range namedInstances(node.FindChildren("application-set")) {
-		as := &ApplicationSet{Name: inst.name}
-
-		for _, member := range inst.node.Children {
-			// An application-set member is either an individual application
-			// reference (`application <name>`) or a nested application-set
-			// reference (`application-set <name>`). Both kinds are stored in
-			// as.Applications; ExpandApplicationSet distinguishes them by
-			// looking each member name up in apps.ApplicationSets and recursing
-			// (max depth 3). Dropping the nested-set arm here silently lost the
-			// child set's applications from the parent, so a policy matching the
-			// parent set under-matched (#2068). This mirrors compileAddressBook,
-			// which handles both `address` and `address-set` members.
-			switch member.Name() {
-			case "application", "application-set":
-				v := nodeVal(member)
-				if v != "" {
-					as.Applications = append(as.Applications, v)
-				}
-			}
-		}
-
-		apps.ApplicationSets[as.Name] = as
-	}
-
-	return nil
-}
-
-// parseApplicationTerms parses an inline term like:
-// "term-name [alg ssh] protocol tcp [source-port 22] [destination-port 22] [inactivity-timeout 86400]"
-// When multiple protocol values are present, returns one Application per
-// unique protocol (each sharing the same ports/timeout/alg).
-func parseApplicationTerms(parentName string, keys []string) []*Application {
-	if len(keys) == 0 {
-		return nil
-	}
-	termName := keys[0]
-
-	var protocols []string
-	var dstPort, srcPort, alg string
-	var timeout int
-
-	for i := 1; i < len(keys); i++ {
-		switch keys[i] {
-		case "protocol":
-			if i+1 < len(keys) {
-				i++
-				protocols = append(protocols, normalizeProtocol(keys[i]))
-			}
-		case "destination-port":
-			if i+1 < len(keys) {
-				i++
-				dstPort = keys[i]
-			}
-		case "source-port":
-			if i+1 < len(keys) {
-				i++
-				srcPort = keys[i]
-			}
-		case "inactivity-timeout", "timeout":
-			if i+1 < len(keys) {
-				i++
-				if v, err := strconv.Atoi(keys[i]); err == nil {
-					timeout = v
-				}
-			}
-		case "alg":
-			if i+1 < len(keys) {
-				i++
-				alg = keys[i]
-			}
-		}
-	}
-
-	// Deduplicate protocols (e.g. "junos-icmp-all" and "icmp" both normalize to "icmp")
-	if len(protocols) == 0 {
-		protocols = []string{""}
-	}
-	seen := make(map[string]bool)
-	var unique []string
-	for _, p := range protocols {
-		if !seen[p] {
-			seen[p] = true
-			unique = append(unique, p)
-		}
-	}
-
-	var result []*Application
-	for _, proto := range unique {
-		name := parentName + "-" + termName
-		if len(unique) > 1 {
-			suffix := proto
-			if suffix == "" {
-				suffix = "any"
-			}
-			name = parentName + "-" + termName + "-" + suffix
-		}
-		result = append(result, &Application{
-			Name:              name,
-			Protocol:          proto,
-			DestinationPort:   dstPort,
-			SourcePort:        srcPort,
-			InactivityTimeout: timeout,
-			ALG:               alg,
-		})
-	}
-	return result
-}
-
-// normalizeProtocol maps Junos protocol aliases to canonical names
-// so that "junos-icmp-all" and "icmp" deduplicate correctly.
-func normalizeProtocol(name string) string {
-	switch strings.ToLower(name) {
-	case "junos-icmp-all", "junos-ping":
-		return "icmp"
-	case "junos-icmp6-all", "junos-pingv6", "icmp6":
-		return "icmpv6"
-	case "junos-gre":
-		return "gre"
-	case "junos-ospf":
-		return "89"
-	case "junos-tcp-any":
-		return "tcp"
-	case "junos-udp-any":
-		return "udp"
-	case "junos-ip-in-ip", "junos-ipip":
-		return "4"
-	default:
-		return name
-	}
-}
-
-// validatePortSpec checks that a port specification is valid.
-// Valid formats: "80", "8080-8090", named ports like "http".
-func validatePortSpec(spec string) error {
-	if spec == "" {
-		return nil
-	}
-	namedPorts := map[string]bool{
-		"http": true, "https": true, "ssh": true, "telnet": true,
-		"ftp": true, "ftp-data": true, "smtp": true, "dns": true,
-		"pop3": true, "imap": true, "snmp": true, "ntp": true,
-		"bgp": true, "ldap": true, "syslog": true,
-	}
-	if namedPorts[strings.ToLower(spec)] {
-		return nil
-	}
-	if strings.Contains(spec, "-") {
-		parts := strings.SplitN(spec, "-", 2)
-		lo, err1 := strconv.Atoi(parts[0])
-		hi, err2 := strconv.Atoi(parts[1])
-		if err1 != nil || err2 != nil {
-			return fmt.Errorf("invalid port range %q: non-numeric", spec)
-		}
-		if lo < 1 || lo > 65535 {
-			return fmt.Errorf("invalid port %d: must be 1-65535", lo)
-		}
-		if hi < 1 || hi > 65535 {
-			return fmt.Errorf("invalid port %d: must be 1-65535", hi)
-		}
-		if lo > hi {
-			return fmt.Errorf("invalid port range %q: start > end", spec)
-		}
-		return nil
-	}
-	port, err := strconv.Atoi(spec)
-	if err != nil {
-		return fmt.Errorf("invalid port %q: not a number or known service", spec)
-	}
-	if port < 1 || port > 65535 {
-		return fmt.Errorf("invalid port %d: must be 1-65535", port)
-	}
-	return nil
-}
-
-// validateProtocol checks that a protocol specification is valid.
-func validateProtocol(proto string) error {
-	validProtos := map[string]bool{
-		"tcp": true, "udp": true, "icmp": true, "icmp6": true, "icmpv6": true,
-		"ospf": true, "gre": true, "ipip": true, "ah": true, "esp": true,
-		"igmp": true, "pim": true, "sctp": true, "vrrp": true, "egp": true,
-	}
-	if validProtos[strings.ToLower(proto)] {
-		return nil
-	}
-	// Accept junos-* protocol aliases
-	if strings.HasPrefix(strings.ToLower(proto), "junos-") {
-		return nil
-	}
-	n, err := strconv.Atoi(proto)
-	if err != nil {
-		return fmt.Errorf("invalid protocol %q", proto)
-	}
-	if n < 0 || n > 255 {
-		return fmt.Errorf("invalid protocol number %d: must be 0-255", n)
-	}
-	return nil
-}
-
-func nodeVal(n *Node) string {
-	if len(n.Keys) >= 2 {
-		return n.Keys[1]
-	}
-	if len(n.Children) > 0 {
-		return n.Children[0].Name()
-	}
-	return ""
 }
